requirements.txt
python-nmap
colorama
pyfiglet
tqdm
